完成基本chat请求

+++ src/main/java/com/rico/rbi/model/entity/History.java
@@ -44,7 +44,7 @@ public class History implements Serializable {
     /**
      * 
      */
-    private Integer modelId;
+    private long modelId;
 
     /**
      *
